pyproject.toml
[project]
name = "program-nova"
version = "0.1.0"
description = "Program Nova - Execution Engine + Dashboard for hierarchical task execution"
readme = "program_nova/README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Eric Magliarditi"}
]

dependencies = [
    "pytest>=7.4.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pyyaml>=6.0.1",
    "httpx>=0.25.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["program_nova"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"

[tool.hatch.build.targets.wheel]
packages = ["program_nova"]

[project.scripts]
nova = "program_nova.cli:main"
